refactor(travel): split 900-line TravelSearchForm into folder module
Pure internal reorganization — no behavior, no prop API, no external
import path changes. Split the single file into 9 focused files
under src/components/travel/TravelSearchForm/:

  index.ts                — barrel re-export (+ "use client" guard)
  TravelSearchForm.tsx    — root component (317 LOC)
  types.ts                — interfaces + RADIUS_META/RADIUS_OPTIONS
  helpers.ts              — 5 pure leg helpers
  LegRow.tsx              — memoized per-leg row + local RadiusPicker
  GhostLegRow.tsx         — "+ ADD LEG" row (authed)
  SignInToAddLegRow.tsx   — sign-in gate (anonymous)
  CompactPill.tsx         — collapsed compact-variant pill
  SubmitBarSummary.tsx    — submit-bar stat line
  BoardingStamp.tsx       — rotated LEG/GHOST/REQUIRED badge

External imports unchanged:
  @/components/travel/TravelSearchForm → resolves through barrel.

Pre-PR review loop:
  - /simplify: no changes needed, split clean.
  - /codex:adversarial-review flagged a real Next.js RSC-boundary
    hazard — re-exporting a "use client" component through a barrel
    without marking the barrel. Fixed by adding "use client" to
    index.ts and confirming `npx next build` compiles /travel clean.
