+++ src/components/travel/TravelSearchForm/index.ts
@@ -1,0 +1,3 @@
+"use client";
+
+export { TravelSearchForm } from "./TravelSearchForm";
